RFDemo/ui/requirements.txt
robotframework-seleniumlibrary~=5.1.3
robotframework-selenium2library~=3.0.0
robotframework-reportportal~=5.2.0
PyUserInput
pyperclip~=1.8.2
configparser~=5.2.0
pywin32
openpyxl~=3.0.9
pykeyboard~=0.1.2
PyMouse~=1.0
robotframework~=4.1.2
Pillow~=9.0.0
bs4~=0.0.1
pytz~=2021.3
lxml~=4.6.4
Crypto~=1.4.1
requests~=2.27.1
requests_toolbelt~=0.9.1
robotframework-requests~=0.9.2
robotframework-pabot~=2.3.1
jsonpath~=0.82